fixed mentioned problems in the review
